fix(conversations): keep the guest password row in sync with allow guests
The password-protection row stayed visible after turning guests off,
since showPasswordProtection was only ever recomputed on a full state
refresh and the optimistic allowGuests() update never touched it. It
now toggles with guestsAllowed like the rest of that state does.

While here: lock the row (checked but disabled, with an explanatory
subtitle) instead of accepting a tap and rejecting it with a snackbar,
now that a server-enforced password can't be removed while guests are
allowed - matching how the sensitive-conversation switch is locked
elsewhere on this screen. The now-unreachable snackbar branch is
removed accordingly.
