+++ app/src/test/java/com/nextcloud/talk/conversationinfo/viewmodel/ConversationInfoViewModelTest.kt
@@ -9,11 +9,9 @@ package com.nextcloud.talk.conversationinfo.viewmodel
 
 import com.nextcloud.talk.api.NcApi
 import com.nextcloud.talk.chat.data.network.ChatNetworkDataSource
-import com.nextcloud.talk.conversationinfo.model.ParticipantModel
 import com.nextcloud.talk.data.user.model.User
 import com.nextcloud.talk.models.json.capabilities.Capabilities
 import com.nextcloud.talk.models.json.capabilities.SpreedCapability
-import com.nextcloud.talk.models.json.participants.Participant
 import com.nextcloud.talk.repositories.passwordpolicy.PasswordPolicyRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Dispatchers
@@ -79,6 +77,21 @@ class ConversationInfoViewModelTest {
         }
 
     @Test
+    fun `turning guests off hides the password protection entry immediately`() =
+        runTest(dispatcher) {
+            val repository = FakeConversationsRepository()
+            val model = viewModel(repository)
+
+            model.allowGuests(user, "token", true, "hunter2")
+            dispatcher.scheduler.advanceUntilIdle()
+            assertEquals(true, model.uiState.value.showPasswordProtection)
+
+            model.allowGuests(user, "token", false)
+
+            assertFalse(model.uiState.value.showPasswordProtection)
+        }
+
+    @Test
     fun `re-enabling guests without a password fails against an enforcing server and reverts the switch`() =
         runTest(dispatcher) {
             val repository = FakeConversationsRepository().apply { failAllowGuests = true }
@@ -104,71 +117,5 @@ class ConversationInfoViewModelTest {
         val result = ConversationInfoViewModel.createConversationNameByParticipants(original, all)
 
         assertEquals(expectedName, result)
-    }
-
-    private fun model(
-        displayName: String,
-        type: Participant.ParticipantType,
-        isOnline: Boolean = true,
-        actorType: Participant.ActorType = Participant.ActorType.USERS
-    ) = ParticipantModel(
-        participant = Participant(
-            actorType = actorType,
-            actorId = displayName.lowercase(),
-            displayName = displayName,
-            type = type
-        ),
-        isOnline = isOnline
-    )
-
-    private fun sortedNames(vararg items: ParticipantModel) =
-        items.sortedWith(ConversationInfoViewModel.PARTICIPANT_COMPARATOR).map { it.participant.displayName }
-
-    @Test
-    fun `owners sort above moderators despite the display name`() {
-        val owner = model("Zoe", Participant.ParticipantType.OWNER)
-        val moderator = model("Alice", Participant.ParticipantType.MODERATOR)
-
-        assertEquals(listOf("Zoe", "Alice"), sortedNames(moderator, owner))
-    }
-
-    @Test
-    fun `owners sort above guest moderators`() {
-        val owner = model("Zoe", Participant.ParticipantType.OWNER)
-        val guestModerator = model("Alice", Participant.ParticipantType.GUEST_MODERATOR)
-
-        assertEquals(listOf("Zoe", "Alice"), sortedNames(guestModerator, owner))
-    }
-
-    @Test
-    fun `moderators sort above plain users`() {
-        val moderator = model("Zoe", Participant.ParticipantType.MODERATOR)
-        val user = model("Alice", Participant.ParticipantType.USER)
-
-        assertEquals(listOf("Zoe", "Alice"), sortedNames(user, moderator))
-    }
-
-    @Test
-    fun `offline participants sort below online ones regardless of rank`() {
-        val offlineOwner = model("Zoe", Participant.ParticipantType.OWNER, isOnline = false)
-        val onlineUser = model("Alice", Participant.ParticipantType.USER)
-
-        assertEquals(listOf("Alice", "Zoe"), sortedNames(offlineOwner, onlineUser))
-    }
-
-    @Test
-    fun `groups and teams sort last`() {
-        val group = model("Aaa Team", Participant.ParticipantType.USER, actorType = Participant.ActorType.GROUPS)
-        val user = model("Zoe", Participant.ParticipantType.USER)
-
-        assertEquals(listOf("Zoe", "Aaa Team"), sortedNames(group, user))
-    }
-
-    @Test
-    fun `same rank falls back to the display name`() {
-        val bob = model("Bob", Participant.ParticipantType.USER)
-        val alice = model("alice", Participant.ParticipantType.USER)
-
-        assertEquals(listOf("alice", "Bob"), sortedNames(bob, alice))
     }
 }
